PZEM groups organized and some other addons
There are hardware limitations with PZEM004T and serial port in Arduino. Not all address can be retrieved with one single port and that's the reason because there are groups of four addresses in each serial port.
Air conditioning machines group have been created also, but because Serial1 is reserved for raspberry comm then I used SoftwareSerial instead.